# libs/data-access/openapi.json
{
"openapi": "3.0.0",
"paths": {
...
"/projects/{projectId}/boards": {
"post": {
"operationId": "BoardsController_create",
"parameters": [
{
"name": "projectId",
"required": true,
"in": "path",
"schema": {
"type": "number"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/CreateBoardDto"
}
}
}
},
"responses": {
"200": {
"description": "",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/BoardResponseDto"
}
}
}
}
},
"security": [
{
"bearer": []
}
],
"tags": [
"Boards"
]
},
"get": {
"operationId": "BoardsController_findAll",
"parameters": [
{
"name": "projectId",
"required": true,
"in": "path",
"schema": {
"type": "number"
}
}
],
"responses": {
"200": {
"description": "",
"content": {
"application/json": {
"schema": {
"type": "array",
"items": {
"$ref": "#/components/schemas/BoardResponseDto"
}
}
}
}
}
},
"security": [
{
"bearer": []
}
],
"tags": [
"Boards"
]
}
},
"/boards/{id}": {
"get": {
"operationId": "BoardsController_findOne",
"parameters": [
{
"name": "id",
"required": true,
"in": "path",
"schema": {
"type": "number"
}
}
],
"responses": {
"200": {
"description": "",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/BoardResponseDto"
}
}
}
}
},
"security": [
{
"bearer": []
}
],
"tags": [
"Boards"
]
},
"put": {
"operationId": "BoardsController_update",
"parameters": [
{
"name": "id",
"required": true,
"in": "path",
"schema": {
"type": "number"
}
}
],
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/UpdateBoardDto"
}
}
}
},
"responses": {
"200": {
"description": "",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/BoardResponseDto"
}
}
}
}
},
"security": [
{
"bearer": []
}
],
"tags": [
"Boards"
]
},
"delete": {
"operationId": "BoardsController_remove",
"parameters": [
{
"name": "id",
"required": true,
"in": "path",
"schema": {
"type": "number"
}
}
],
"responses": {
"200": {
"description": "",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/BoardResponseDto"
}
}
}
}
},
"security": [
{
"bearer": []
}
],
"tags": [
"Boards"
]
}
},
...
},
"info": {
"title": "Flow Desk API",
"description": "Backend Jira clone",
"version": "1.0",
"contact": {}
},
...
"servers": [],
"components": {
"securitySchemes": {
"bearer": {
"scheme": "bearer",
"bearerFormat": "JWT",
"type": "http"
}
},
"schemas": {
...
"CreateBoardDto": {
"type": "object",
"properties": {
"name": {
"type": "string"
}
},
"required": [
"name"
]
},
"BoardResponseDto": {
"type": "object",
"properties": {
"id": {
"type": "number"
},
"name": {
"type": "string"
},
"projectId": {
"type": "number"
},
"createdAt": {
"format": "date-time",
"type": "string"
},
"updatedAt": {
"format": "date-time",
"type": "string"
}
},
"required": [
"id",
"name",
"projectId",
"createdAt",
"updatedAt"
]
},
"UpdateBoardDto": {
"type": "object",
"properties": {
"name": {
"type": "string"
}
}
},
...
}
}
}